Implemented mutation.
I added mutation where the baby of parents with different nesting time preferences has a small chance of flipping to the other preference. Unfortunately this is *so* small that the odds of it happening with the population sizes they're actually working with are really epsilon, so we may have to drop the whole binary nesting time assumption and go back to the continuously varying nesting time idea. That's entirely do-able, but will require a fair bit of new coding.