Replaced the "demo" simulation with calls to the fleshed out simulation.
This replaces all the demo computations with the construction of "real" populations and simulators, and calls to update those.

It seems to work fine, although I suspect there's some sort of O(N^2) behavior, as it definitely seems to get slower over time. This may be in the `@values` array, where we keep pushing new values onto the end, which is likely to be O(N^2) in the long run.