Added support for downloading simulation data.
There's now a "Download" button that extracts the data from the most recent collection of runs and downloads it as a CSV file. If you're using a recent version of Chrome or Firefox it actually initiates the download. If you're using Safari it opens a new window/tab containing the data. I have no idea what happens with IE.

At the moment the parameters used for that simulation are _not_ being saved, but they probably should be. Not sure how to best do that, however. We may take the R approach and just have columns for them with tons of repeated data.