Fix #6: Collect 2 eggs instead of 1
I had us only collecting one egg, when apparently both eggs are collected during forced renesting, and
both of those eggs typically hatch and become fledglings.

This fixes that, although in a fairly hack-ish way. I think the “proper” fix would be to model eggs directly (which is
Issue #7) instead of assuming a one-to-one correspondence between nests and eggs.