Set initial population to be of pairing age
I had all the birds in the initial population to have age zero, which means that there was no
nesting/breeding for the first few years, but we were still occasionally losing birds due to the
mortality passes. This means runs always started with a slight population dip before things
started to grow (if the configuration supported growth).

This commit fixes that by setting the initial age of the birds to be one more than the pairing age,
allowing simulations to “hit the ground running”.