Convert `collectionProbability` to a “variable”
The set up in `ModelParameters` has all the parameters as essentially “constants”, but
we want the `collectionProbability` parameter to change over time now, so I created
get/set functions for it. This required making a bucket of tedious changes elsewhere
(especially in the tests) to use the new function instead of directly referring to the
variable.